# Conflicts:
#	src/main/java/Controllers/BudgetModelController.java
#	src/main/resources/overview2.fxml
#	target/classes/Controllers/BudgetModelController.class
#	target/classes/Model/BudgetModel.class
#	target/classes/Model/Category.class
#	target/classes/com/example/budgetmaker2_0/cateOVList.fxml
#	target/classes/overview2.fxml

+++ src/main/java/Model/Transaction.java
@@ -1,23 +1,27 @@
 package Model;
 
+import javafx.scene.control.ChoiceBox;
+import javafx.scene.control.DatePicker;
+import javafx.scene.control.Label;
+import javafx.scene.control.TextField;
+
 import java.time.LocalDate;
 
 public class Transaction {
-    private double transactionAmount;
+    private int transactionAmount;
     private String name;
     private String notes;
     private Category category;
     private LocalDate date;
 
 
-    public Transaction(Double amount, String name, String notes, Category category, LocalDate date){
+    public Transaction(int amount, String name, String notes, Category category, LocalDate date){
         this.transactionAmount = amount;
         this.name = name;
         this.notes = notes;
         this.category = category;
         this.date = date;
     }
-
 
     //Getters
     public double getTransactionAmount(){
